Refactor researcher function and update tools
